Extract models in `dataset_urls` to a different file
